health-os/integrations/whoop/requirements.txt
# WHOOP Integration Dependencies
whoopy>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.28.0
